add game.py
とりあえずhnb 押せるところまで作った
細かい仕様決めてから作ったほうが良さそう

+++ main.py
@@ -10,11 +10,13 @@ from dotenv import load_dotenv
 load_dotenv()
 
 import datetime
+from zoneinfo import ZoneInfo
 import json
 import random
 
 from cogs.role import *
 from cogs.miscellaneous import miscmenu_view, miscmenu_page
+from cogs.game import gamemenuview, embedbox_game
 # ---------------------------------------
 
 bot_prefix = "!"
@@ -29,7 +31,8 @@ bot = commands.Bot(command_prefix=bot_prefix, intents=intents)
 
 coglist = [
     'cogs.role',
-    'cogs.miscellaneous'
+    'cogs.miscellaneous',
+    'cogs.game'
 ]
 
 # ---------------------------------------
@@ -65,9 +68,9 @@ async def menu(ctx,isOnly=None):
     e_page = []
     v_page = []
     embed.add_field(name="!role",value="ロールに関するメニュー",inline=True)
+    embed.add_field(name="!game",value="ゲームメニュー",inline=True)
     embed.add_field(name="!misc",value="その他、細かいもの",inline=True)
-    embed.add_field(name="name3",value="value3",inline=True)
-    
+
     if isOnly =="1":
         embed.set_footer(text=(f"{author_name}のみ操作可能"),icon_url=author_image)
         e_page.append(embed)
@@ -85,11 +88,9 @@ class menu_button(discord.ui.View):
         super().__init__(timeout=timeout)
         self.author = author
         self.isOnly = isOnly
-        # self.isTree = isTree
         self.e_page = e_page
         self.v_page = v_page
         v_page.append(self)
-        # e_page.append(embed)
 
     async def interaction_check(self, interaction: discord.Interaction, /) -> bool:
         try:
@@ -133,44 +134,30 @@ class menu_button(discord.ui.View):
         await interaction.response.edit_message(embed=Embeds,view = Views)
 
     @discord.ui.button(
+        label=(f'!game'),
+        style=discord.ButtonStyle.primary,
+    )
+
+    async def gotogame(self,interaction: discord.Interaction,button: discord.ui.Button):
+        author = interaction.user
+        evs = embedbox_game(author,self.isOnly)
+        Embeds = evs.e_game_top()
+        self.e_page.append(Embeds)
+        Views = gamemenuview(isOnly=self.isOnly,author=author,e_page=self.e_page,v_page=self.v_page)
+        await interaction.response.edit_message(embed=Embeds,view=Views)
+
+    @discord.ui.button(
         label=(f'もっかい同じ送る(テスト)'),
         style=discord.ButtonStyle.primary,
     )
 
     async def aho(self, interaction: discord.Interaction,button: discord.ui.Button):
         await interaction.response.send_message(embed=self.e_page[-1],view=self.v_page[-1])
-        
 
 @bot.command()
 async def register(ctx):
-    guild :discord.Guild= ctx.guild
-    json_data ={
-    "name": guild.name,
-    "id" : guild.id,
-    "registerdate": str(datetime.datetime.now()),
-    "role":[]
-    }
-    path = "./bot_witch/guilds/" + str(guild.id) + ".json"
-    if (os.path.isfile(path)):
-        embeds = discord.Embed(color=0x880088,title="すでに登録済みです",description="一部の保存されているデータが消去される可能性があります。\n本当に変更しますか？")
-        await ctx.send(embed=embeds,view=registerconfirm())
-        return
-    with open(path, "a+") as file:
-        json.dump(json_data,file,indent=4)    # 追加する値を入力
-    await ctx.send("登録完了しました")
-
-    # with open("sample.json", "w") as f:
-    #     json.dump(json_data, f)
-class registerconfirm(discord.ui.View):
-    def __init__(self,*,timeout=None) -> None:
-        super().__init__(timeout=timeout)
-    
-    @discord.ui.button(
-    label=(f'変更する'),
-    style=discord.ButtonStyle.green,)
-
-    async def callback(self, interaction: discord.Interaction,button: discord.ui.Button):
-        guild = interaction.guild
+    if ctx.guild:
+        guild :discord.Guild= ctx.guild
         json_data ={
         "name": guild.name,
         "id" : guild.id,
@@ -178,10 +165,77 @@ class registerconfirm(discord.ui.View):
         "role":[]
         }
         path = "./bot_witch/guilds/" + str(guild.id) + ".json"
+        if (os.path.isfile(path)):
+            embeds = discord.Embed(color=0x880088,title="すでに登録済みです",description="一部の保存されているデータが消去される可能性があります。\n本当に変更しますか？")
+            await ctx.send(embed=embeds,view=registerconfirm())
+            return
+        with open(path, "a+") as file:
+            json.dump(json_data,file,indent=4)    # 追加する値を入力
+        await ctx.send("登録完了しました")
+        return
+    else: #DMやったとき
+        user = ctx.author
+        tz = ZoneInfo('Asia/Tokyo')
+        timedate = datetime.datetime.now(tz=tz)
+        result = timedate.strftime('%Y/%m/%d (%Z)')
+        json_data ={
+        "name": user.name,
+        "id" : user.id,
+        "registerdate": result,
+        "battle":"0",
+        "win"   :"0",
+        "draw"  :"0"
+        }
+        path = "./bot_witch/users/" + str(user.id) + ".json"
+        if (os.path.isfile(path)):
+            embeds = discord.Embed(color=0x880088,title="すでに登録済みです",description="一部の保存されているデータが消去される可能性があります。\n本当に変更しますか？")
+            await ctx.send(embed=embeds,view=registerconfirm(isdm = True))
+            return
         with open(path, "w") as file:
             json.dump(json_data,file,indent=4)    # 追加する値を入力
-        embed = discord.Embed(color=0x00ff00,description="登録完了しました")
-        await interaction.response.edit_message(embed=embed,view=None)
+        await ctx.send("登録完了しました")
+        return
+
+class registerconfirm(discord.ui.View):
+    def __init__(self,*,timeout=None,isdm:bool = False) -> None:
+        super().__init__(timeout=timeout)
+        self.isdm = isdm
+    @discord.ui.button(
+    label=(f'変更する'),
+    style=discord.ButtonStyle.green,)
+
+    async def callback(self, interaction: discord.Interaction,button: discord.ui.Button):
+        if self.isdm:
+            user = interaction.user
+            tz = ZoneInfo('Asia/Tokyo')
+            timedate = datetime.datetime.now(tz=tz)
+            result = timedate.strftime('%Y/%m/%d (%Z)')
+            json_data ={
+            "name": user.name,
+            "id" : user.id,
+            "registerdate": result
+            }
+            path = "./bot_witch/users/" + str(user.id) + ".json"
+            with open(path, "w") as file:
+                json.dump(json_data,file,indent=4)    # 追加する値を入力
+            embed = discord.Embed(color=0x00ff00,description="登録完了しました")
+            await interaction.response.edit_message(embed=embed,view=None)
+            return
+        else:
+            guild = interaction.guild
+            json_data ={
+            "name": guild.name,
+            "id" : guild.id,
+            "registerdate": str(datetime.datetime.now()),
+            "role":[]
+            }
+            path = "./bot_witch/guilds/" + str(guild.id) + ".json"
+            with open(path, "w") as file:
+                json.dump(json_data,file,indent=4)    # 追加する値を入力
+            embed = discord.Embed(color=0x00ff00,description="登録完了しました")
+            await interaction.response.edit_message(embed=embed,view=None)
+            return
+
 
     @discord.ui.button(
     label=(f'変更しない'),
